chore(data-lookup): address framework audit findings
Follow-up to PR #1124. /audit flagged 7 findings across 4 categories; this PR
resolves all of them.

Metrics naming (1):
- Rename granit.datalookup.* -> granit.data_lookup.* to match the snake-case
  module convention used by sibling packages (granit.query_engine.*,
  granit.data_exchange.*). Meter name stays PascalCase Granit.DataLookup.

File organization (3):
- Split ILookupRegistry.cs: move LookupManifestEntry to its own file.
- Split LookupRegistry.cs: move IKindProviderLookupSource to its own file.
- Split LookupManifestResponse.cs: move LookupManifestEntryResponse to its own
  file. Matches the architecture-test rule "one primary type per file".

OpenAPI metadata (2):
- Remove redundant .ProducesProblem(403) on manifest and resolve endpoints --
  auto-injected by ProblemDetailsResponseOperationTransformer when [Authorize]
  is present.
- Add a legitimate 400 path in ResolveAsync for empty 'value' query parameter;
  the ProblemHttpResult branch of the Results<...> union now has a matching
  handler path (no cosmetic branch). Extension emits .ProducesProblem(400).

Documentation (1):
- New docs-site page docs/dotnet/business/data-lookup.mdx covering why,
  package layout, canonical shapes, registration patterns (Enum / IQueryable /
  custom), endpoints, auth, Empty Scope Trap mitigation, and diagnostics.

Contract documentation (1):
- Clarify in ILookupSource XML docs that authorization AND scope-key validation
  are cross-cutting concerns enforced by the endpoint layer. Programmatic
  consumers bypassing the endpoints must apply both checks themselves before
  invoking SearchAsync.
