+++ .mcp-code-index.json
@@ -12376,8 +12376,8 @@
       "fqn": "Granit.DataLookup.Endpoints.Dtos.LookupManifestEntryResponse",
       "kind": "record",
       "project": "Granit.DataLookup.Endpoints",
-      "file": "src/Granit.DataLookup.Endpoints/Dtos/LookupManifestResponse.cs",
-      "line": 14,
+      "file": "src/Granit.DataLookup.Endpoints/Dtos/LookupManifestEntryResponse.cs",
+      "line": 10,
       "members": []
     },
     {
@@ -12386,7 +12386,7 @@
       "kind": "record",
       "project": "Granit.DataLookup.Endpoints",
       "file": "src/Granit.DataLookup.Endpoints/Dtos/LookupManifestResponse.cs",
-      "line": 7,
+      "line": 5,
       "members": []
     },
     {
@@ -12553,7 +12553,7 @@
       "kind": "interface",
       "project": "Granit.DataLookup.Abstractions",
       "file": "src/Granit.DataLookup.Abstractions/Registry/ILookupRegistry.cs",
-      "line": 15,
+      "line": 14,
       "members": [
         {
           "name": "Resolve",
@@ -12574,8 +12574,8 @@
       "fqn": "Granit.DataLookup.Registry.LookupManifestEntry",
       "kind": "record",
       "project": "Granit.DataLookup.Abstractions",
-      "file": "src/Granit.DataLookup.Abstractions/Registry/ILookupRegistry.cs",
-      "line": 34,
+      "file": "src/Granit.DataLookup.Abstractions/Registry/LookupManifestEntry.cs",
+      "line": 13,
       "members": []
     },
     {
@@ -12600,7 +12600,7 @@
       "kind": "interface",
       "project": "Granit.DataLookup.Abstractions",
       "file": "src/Granit.DataLookup.Abstractions/Sources/ILookupSource.cs",
-      "line": 32,
+      "line": 45,
       "members": [
         {
           "name": "SearchAsync",
